Adding new epicollection view and edit pages. Modifying existing edit and view collection pages to become a module
